fishow_django/requirements.txt
asgiref==3.2.4
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.44.0
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-email-users==1.1.0
django-filter==2.4.0
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.0.0
Pygments==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
pytils==0.3
django-thumbnails==0.4.0
